fey-tex at last!
The eleven pages of the fey plot have so-far taken twenty-one hours.
Three and a half of those hours consisted of converting the markdown
writing to LaTeX.

It looks like the benefits of straight-forward markdown writing aren't
worth it, although the process could be a good deal faster with a
well-knapped script.