ensure daily bailey has a back page
The bailey market is the first page of the minizine introductions.  It's
the only page that's in A4 format.  It needs to have a back-page so that
people can easily orient the pages when printing (as long as the bailey
market handout is the right way round, the rest will be too).